Working on spliiting chunks in webpack
